Add Get Started guide and environment setup instructions to README
These updates to the README include a step-by-step guide to assist new users in getting started with the application. Additionally, an in-depth explanation of setting up requisite environment variables has been added to streamline the setup process, along with instructions for publishing the Prisma Schema and updating NextAuth.js callback URLs.